Overline of the memory management
As discuted with Quentin, this commit introduce the ouline of the memory
management system we are designing :

Two hashmap, one for the scope another one as an unorganized data store.
The first one has references to the second one.

For the moment the stack is not implementeted, hence every variables,
classes, attributes or methods are available everywhere like globals.
This will change soon once we have implemented more stuff.